Titanic: Machine Learning from Disaster
The sinking of the Titanic is one of the most infamous shipwrecks in history. On April 15, 1912, during her maiden voyage, the Titanic sank after colliding with an iceberg, killing 1502 out of 2224 passengers and crew. That's why the name Dianic. This is a very unforgettable disaster that no one in the world can forget.

It took about $7.5 million to build the Titanic and it sunk under the ocean due to a collision. The Titanic Dataset is a very good dataset for beginners to start a journey in data science and participate in competitions in Kaggle.